Move updateConfig logic into config.go
This function will be required in the upcoming cert rotation logic and
is currently ifdef'd into code via vpn.go.
