+++ internal/vpn.go
@@ -14,18 +14,6 @@ import (
 	"strings"
 )
 
-type ConfigFileReq struct {
-	Name        string   `json:"name"`
-	Value       string   `json:"value"`
-	Unencrypted bool     `json:"unencrypted"`
-	OnChanged   []string `json:"on-changed,omitempty"`
-}
-
-type ConfigCreateRequest struct {
-	Reason string          `json:"reason"`
-	Files  []ConfigFileReq `json:"files"`
-}
-
 // Create a private key and return the derived public key
 func generateKey(privKeyPath string) (string, error) {
 	pkey, err := exec.Command("wg", "genkey").Output()
@@ -53,51 +41,6 @@ func generateKey(privKeyPath string) (string, error) {
 	}
 	return strings.TrimSpace(string(pub)), nil
 
-}
-
-func updateConfig(app *App, client *http.Client, pubkey string) error {
-	updated := ""
-	content, err := ioutil.ReadFile(filepath.Join(app.SecretsDir, "wireguard-client"))
-	if err != nil {
-		if os.IsNotExist(err) {
-			updated = "enabled=0\n" // This isn't enabled
-		} else {
-			return err
-		}
-	}
-	written := false
-
-	for _, line := range strings.Split(string(content), "\n") {
-		if strings.HasPrefix(line, "pubkey=") {
-			updated += "pubkey=" + pubkey + "\n"
-			written = true
-		} else {
-			updated += line + "\n"
-		}
-	}
-	if !written {
-		updated += "pubkey=" + pubkey + "\n"
-	}
-	updated = strings.TrimSpace(updated)
-
-	ccr := ConfigCreateRequest{
-		Reason: "Set Wireguard pubkey from fioconfig",
-		Files: []ConfigFileReq{
-			ConfigFileReq{
-				Name:        "wireguard-client",
-				Unencrypted: true,
-				Value:       updated,
-			},
-		},
-	}
-	res, err := httpPatch(client, app.configUrl, ccr)
-	if err != nil {
-		return err
-	}
-	if res.StatusCode != 201 {
-		return fmt.Errorf("Unable to update: %s - HTTP_%d", app.configUrl, res.StatusCode, string(res.Body))
-	}
-	return nil
 }
 
 func vpnBugFix(app *App, sotaConfig string) bool {
